Syncing SOS with my repo

+++ src/symmetric_heap_c.c
@@ -4,7 +4,7 @@
  * DE-AC04-94AL85000 with Sandia Corporation, the U.S.  Government
  * retains certain rights in this software.
  *
- * Copyright (c) 2016 Intel Corporation. All rights reserved.
+ * Copyright (c) 2017 Intel Corporation. All rights reserved.
  * This software is available to you under the BSD license.
  *
  * This file is part of the Sandia OpenSHMEM software package. For license
@@ -40,6 +40,9 @@
 #pragma weak shmem_malloc = pshmem_malloc
 #define shmem_malloc pshmem_malloc
 
+#pragma weak shmem_calloc = pshmem_calloc
+#define shmem_calloc pshmem_calloc
+
 #pragma weak shmem_align = pshmem_align
 #define shmem_align pshmem_align
 
@@ -61,15 +64,19 @@
 #pragma weak shfree = pshfree
 #define shfree pshfree
 
+#pragma weak shmemx_malloc_with_hints = pshmemx_malloc_with_hints
+#define shmemx_malloc_with_hints pshmemx_malloc_with_hints
+
 #endif /* ENABLE_PROFILING */
 
 static char *shmem_internal_heap_curr = NULL;
-static int shmem_internal_use_malloc = 0;
 
 void* dlmalloc(size_t);
+void* dlcalloc(size_t, size_t);
 void  dlfree(void*);
 void* dlrealloc(void*, size_t);
 void* dlmemalign(size_t, size_t);
+
 
 /*
  * scan /proc/mounts for a huge page file system with the
@@ -78,10 +85,10 @@ void* dlmemalign(size_t, size_t);
  * On success return 0, else -1.
  */
 
+#ifdef __linux__
 static int find_hugepage_dir(size_t page_size, char **directory)
 {
     int ret = -1;
-#ifdef __linux__
     struct statfs pg_size;
     struct mntent *mntent;
     FILE *fd;
@@ -104,7 +111,7 @@ static int find_hugepage_dir(size_t page_size, char **directory)
 
         path = mntent->mnt_dir;
         if (statfs(path, &pg_size) == 0) {
-            if (pg_size.f_bsize == page_size) {
+            if ((size_t) pg_size.f_bsize == page_size) {
                 *directory = strdup(path);
                 ret = 0;
                 break;
@@ -113,10 +120,9 @@ static int find_hugepage_dir(size_t page_size, char **directory)
     }
 
     endmntent(fd);
-#endif
-
     return ret;
 }
+#endif /* __linux__ */
 
 /* shmalloc and friends are defined to not be thread safe, so this is
    fine.  If they change that definition, this is no longer fine and
@@ -157,28 +163,25 @@ static void *mmap_alloc(size_t bytes)
     char *file_name = NULL;
     int fd = 0;
     char *directory = NULL;
-    const char basename[] = "hugepagefile.SOS";
-    int size;
     void *requested_base =
         (void*) (((unsigned long) shmem_internal_data_base +
                   shmem_internal_data_length + 2 * ONEGIG) & ~(ONEGIG - 1));
     void *ret;
 
-    if (shmem_internal_heap_use_huge_pages) {
+#ifdef __linux__
+    /* huge page support only on Linux for now, default is to use 2MB large pages */
+    if (shmem_internal_params.SYMMETRIC_HEAP_USE_HUGE_PAGES) {
+        const char basename[] = "hugepagefile.SOS";
 
-        /*
-         * check what /proc/mounts has for explicit huge page support
-         */
+        /* check what /proc/mounts has for explicit huge page support */
+        if (find_hugepage_dir(shmem_internal_params.SYMMETRIC_HEAP_PAGE_SIZE,
+                             &directory) == 0)
+        {
+            int size = snprintf(NULL, 0, "%s/%s.%d", directory, basename, getpid());
 
-        if(find_hugepage_dir(shmem_internal_heap_huge_page_size, &directory) == 0) {
-
-            size = snprintf(NULL, 0, "%s/%s.%d", directory, basename, getpid());
             if (size < 0) {
-
-                RAISE_WARN_STR("snprint returned error, cannot use huge pages");
-
+                RAISE_WARN_STR("snprintf returned error, cannot use huge pages");
             } else {
-
                 file_name = malloc(size + 1);
                 if (file_name) {
                     sprintf(file_name, "%s/%s.%d", directory, basename, getpid());
@@ -187,14 +190,14 @@ static void *mmap_alloc(size_t bytes)
                         RAISE_WARN_STR("file open failed, cannot use huge pages");
                         fd = 0;
                     } else {
-                        /* have to round up
-                           by the pagesize being used */
-                        bytes = CEILING(bytes, shmem_internal_heap_huge_page_size);
+                        /* have to round up by the pagesize being used */
+                        bytes = CEILING(bytes, shmem_internal_params.SYMMETRIC_HEAP_PAGE_SIZE);
                     }
                 }
             }
         }
     }
+#endif /* __linux__ */
 
     ret = mmap(requested_base,
                bytes,
@@ -210,7 +213,8 @@ static void *mmap_alloc(size_t bytes)
         ret = NULL;
     }
     if (fd) {
-        unlink(file_name);
+        if (file_name)
+            unlink(file_name);
         close(fd);
     }
     if (directory) {
@@ -224,14 +228,13 @@ static void *mmap_alloc(size_t bytes)
 
 
 int
-shmem_internal_symmetric_init(size_t requested_length, int use_malloc)
+shmem_internal_symmetric_init(void)
 {
-    shmem_internal_use_malloc = use_malloc;
-
     /* add library overhead such that the max can be shmalloc()'ed */
-    shmem_internal_heap_length = requested_length + (1024*1024);
+    shmem_internal_heap_length = shmem_internal_params.SYMMETRIC_SIZE +
+                                 SHMEM_INTERNAL_HEAP_OVERHEAD;
 
-    if (0 == shmem_internal_use_malloc) {
+    if (!shmem_internal_params.SYMMETRIC_HEAP_USE_MALLOC) {
         shmem_internal_heap_base =
             shmem_internal_heap_curr =
             mmap_alloc(shmem_internal_heap_length);
@@ -249,7 +252,7 @@ int
 shmem_internal_symmetric_fini(void)
 {
     if (NULL != shmem_internal_heap_base) {
-        if (0 == shmem_internal_use_malloc) {
+        if (!shmem_internal_params.SYMMETRIC_HEAP_USE_MALLOC) {
             munmap( (void*)shmem_internal_heap_base, (size_t)shmem_internal_heap_length );
         } else {
             free(shmem_internal_heap_base);
@@ -278,9 +281,11 @@ shmem_internal_shmalloc(size_t size)
 void SHMEM_FUNCTION_ATTRIBUTES *
 shmem_malloc(size_t size)
 {
-    void *ret;
+    void *ret = NULL;
 
     SHMEM_ERR_CHECK_INITIALIZED();
+
+    if (size == 0) return ret;
 
     SHMEM_MUTEX_LOCK(shmem_internal_mutex_alloc);
     ret = dlmalloc(size);
@@ -291,6 +296,23 @@ shmem_malloc(size_t size)
     return ret;
 }
 
+void SHMEM_FUNCTION_ATTRIBUTES *
+shmem_calloc(size_t count, size_t size)
+{
+    void *ret = NULL;
+
+    SHMEM_ERR_CHECK_INITIALIZED();
+
+    if (size == 0 || count == 0) return ret;
+
+    SHMEM_MUTEX_LOCK(shmem_internal_mutex_alloc);
+    ret = dlcalloc(count, size);
+    SHMEM_MUTEX_UNLOCK(shmem_internal_mutex_alloc);
+
+    shmem_internal_barrier_all();
+
+    return ret;
+}
 
 void SHMEM_FUNCTION_ATTRIBUTES
 shmem_free(void *ptr)
@@ -302,13 +324,7 @@ shmem_free(void *ptr)
 
     shmem_internal_barrier_all();
 
-    /* It's fine to call dlfree with NULL, but better to avoid unnecessarily
-     * taking the mutex in the threaded case. */
-    if (ptr != NULL) {
-        SHMEM_MUTEX_LOCK(shmem_internal_mutex_alloc);
-        dlfree(ptr);
-        SHMEM_MUTEX_UNLOCK(shmem_internal_mutex_alloc);
-    }
+    shmem_internal_free(ptr);
 }
 
 
@@ -318,6 +334,8 @@ shmem_realloc(void *ptr, size_t size)
     void *ret;
 
     SHMEM_ERR_CHECK_INITIALIZED();
+
+    if (size == 0 && ptr == NULL) return ptr;
     if (ptr != NULL) {
       SHMEM_ERR_CHECK_SYMMETRIC_HEAP(ptr);
     }
@@ -342,10 +360,11 @@ shmem_realloc(void *ptr, size_t size)
 void SHMEM_FUNCTION_ATTRIBUTES *
 shmem_align(size_t alignment, size_t size)
 {
-    void *ret;
+    void *ret = NULL;
 
     SHMEM_ERR_CHECK_INITIALIZED();
 
+    if (size == 0) return ret;
     if (alignment == 0)
         return NULL;
 
@@ -386,4 +405,28 @@ void SHMEM_FUNCTION_ATTRIBUTES * shrealloc(void *ptr, size_t size)
 void SHMEM_FUNCTION_ATTRIBUTES * shmemalign(size_t alignment, size_t size)
 {
     return shmem_align(alignment, size);
+}
+
+
+void SHMEM_FUNCTION_ATTRIBUTES *
+shmemx_malloc_with_hints(size_t size, long hints)
+{
+    void *ret = NULL;
+
+    SHMEM_ERR_CHECK_INITIALIZED();
+
+    if (size == 0) return ret;
+
+    // Check for valid hints
+    if(hints > SHMEMX_MALLOC_MAX_HINTS || hints < 0) {
+        RAISE_WARN_MSG("Ignoring invalid hint for shmem_malloc_with_hints(%ld)\n", hints);
+    }
+
+    SHMEM_MUTEX_LOCK(shmem_internal_mutex_alloc);
+    ret = dlmalloc(size);
+    SHMEM_MUTEX_UNLOCK(shmem_internal_mutex_alloc);
+
+    shmem_internal_barrier_all();
+
+    return ret;
 }
